Replace old modules with new
